Fixed Volume Option Bug
The way the volume adjustment worked before, the game called a method from the GameOptions script which returned the player's chosen volume level, and multiplied that value by the audio sources volume level, or by a default value of 1 if the player didn't make any audio adjustments. However the GameplayScene did not have access to the script, so the value was never created thus causing the sound effect to error out.
Created empty GameOptions object in gameplay scene, added GameOptions script. Fixed the issue.