Added pause time, and fixed coroutine waiting to finish last full second
Pressing 0 will now pause the game, well techincally it doesn't pause it, it sets it to .000001 game minute to real seconds, so it might as well be paused. Did it this way to avoid dividing by 0. It will also no longer wait to finish the last second to move forward with time, otherwise you'd have to wait years for the slow time to update to a faster time. It's now instant.